requirements.txt
flask

flask-sqlalchemy
Flask-WTF
Flask-Login
wtforms[email]

gunicorn

pytest
pytest-flask
pytest-cov